requirements-armv6.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements-armv6.in --output-file requirements-armv6.txt
aiofiles==24.1.0
    # via quart
aiosqlite==0.21.0
    # via -r requirements-armv6.in
alembic==1.16.5
    # via -r requirements-armv6.in
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via httpx
apscheduler==3.11.0
    # via -r requirements-armv6.in
blinker==1.9.0
    # via
    #   flask
    #   quart
brotli==1.1.0
    # via
    #   httpx
    #   quart-compress
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
    #   sentry-sdk
charset-normalizer==3.4.3
    # via requests
click==8.3.0
    # via
    #   ddgs
    #   flask
    #   quart
colorama==0.4.6
    # via griffe
ddgs==9.6.0
    # via pydantic-ai-slim
flask==3.1.2
    # via quart
genai-prices==0.0.29
    # via pydantic-ai-slim
greenlet==3.2.4
    # via sqlalchemy
griffe==1.14.0
    # via pydantic-ai-slim
h11==0.16.0
    # via
    #   httpcore
    #   hypercorn
    #   wsproto
h2==4.3.0
    # via
    #   httpx
    #   hypercorn
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   -r requirements-armv6.in
    #   ddgs
    #   genai-prices
    #   pydantic-ai-slim
    #   pydantic-graph
    #   python-telegram-bot
hypercorn==0.17.3
    # via quart
hyperframe==6.1.0
    # via h2
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==8.7.0
    # via opentelemetry-api
itsdangerous==2.2.0
    # via
    #   flask
    #   quart
jinja-ui-kit @ git+https://github.com/byzantime/jinja-ui-kit.git@125c447e97e58a1065714cd5abef2de1f263357a
    # via -r requirements-armv6.in
jinja2==3.1.6
    # via
    #   flask
    #   jinja-ui-kit
    #   quart
logfire-api==4.10.0
    # via pydantic-graph
lxml==6.0.2
    # via ddgs
mako==1.3.10
    # via alembic
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   mako
    #   quart
    #   werkzeug
numpy==2.3.3
    # via -r requirements-armv6.in
opentelemetry-api==1.37.0
    # via pydantic-ai-slim
primp==0.15.0
    # via ddgs
priority==2.0.0
    # via hypercorn
pydantic==2.11.9
    # via
    #   genai-prices
    #   pydantic-ai-slim
    #   pydantic-graph
pydantic-ai-slim==1.0.10
    # via -r requirements-armv6.in
pydantic-core==2.33.2
    # via pydantic
pydantic-graph==1.0.10
    # via pydantic-ai-slim
pyscss==1.4.0
    # via quart-assets
python-dotenv==1.1.1
    # via -r requirements-armv6.in
python-telegram-bot==22.5
    # via -r requirements-armv6.in
pyyaml==6.0.3
    # via
    #   quart-assets
    #   webassets
quart==0.20.0
    # via
    #   quart-assets
    #   quart-auth
    #   quart-compress
    #   quart-flask-patch
quart-assets==0.1.2
    # via -r requirements-armv6.in
quart-auth==0.11.0
    # via -r requirements-armv6.in
quart-compress==0.2.1
    # via -r requirements-armv6.in
quart-flask-patch==0.3.0
    # via -r requirements-armv6.in
requests==2.32.5
    # via vadersentiment
sentry-sdk==2.39.0
    # via -r requirements-armv6.in
setuptools==80.9.0
    # via zope-dottedname
six==1.17.0
    # via pyscss
sniffio==1.3.1
    # via anyio
socksio==1.0.0
    # via httpx
sqlalchemy==2.0.43
    # via
    #   -r requirements-armv6.in
    #   alembic
tornado==6.5.2
    # via python-telegram-bot
typing-extensions==4.15.0
    # via
    #   aiosqlite
    #   alembic
    #   opentelemetry-api
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-ai-slim
    #   pydantic-graph
tzlocal==5.3.1
    # via apscheduler
urllib3==2.5.0
    # via
    #   requests
    #   sentry-sdk
vadersentiment==3.3.2
    # via -r requirements-armv6.in
webassets==3.0.0
    # via quart-assets
werkzeug==3.1.3
    # via
    #   flask
    #   quart
wsproto==1.2.0
    # via hypercorn
zipp==3.23.0
    # via importlib-metadata
zope-dottedname==7.0
    # via webassets
